app/requirements.txt
azure-ai-contentsafety >= 1.0.0
google-api-core >= 2.15.0
google-auth >= 2.25.2
google-cloud-language >= 2.12.0
azure-cognitiveservices-vision-contentmoderator
python-dotenv
discord
discord.py
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
pymongo
paramiko